Remove StudentSessionData class
