+++ src/main/java/seedu/taassist/logic/commands/AssignCommand.java
@@ -14,8 +14,8 @@ import seedu.taassist.logic.parser.ParserStudentIndexUtil;
 import seedu.taassist.logic.parser.exceptions.ParseException;
 import seedu.taassist.model.Model;
 import seedu.taassist.model.moduleclass.ModuleClass;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.student.Student;
-import seedu.taassist.model.student.StudentModuleData;
 
 /**
  * Assigns students to a class.
