+++ src/main/java/seedu/taassist/logic/parser/AddCommandParser.java
@@ -14,12 +14,12 @@ import java.util.stream.Collectors;
 import seedu.taassist.logic.commands.AddCommand;
 import seedu.taassist.logic.parser.exceptions.ParseException;
 import seedu.taassist.model.moduleclass.ModuleClass;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.student.Address;
 import seedu.taassist.model.student.Email;
 import seedu.taassist.model.student.Name;
 import seedu.taassist.model.student.Phone;
 import seedu.taassist.model.student.Student;
-import seedu.taassist.model.student.StudentModuleData;
 
 /**
  * Parses input arguments and creates a new AddCommand object.
