+++ src/main/java/seedu/taassist/model/util/SampleDataUtil.java
@@ -7,12 +7,12 @@ import java.util.stream.Collectors;
 import seedu.taassist.model.ReadOnlyTaAssist;
 import seedu.taassist.model.TaAssist;
 import seedu.taassist.model.moduleclass.ModuleClass;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.student.Address;
 import seedu.taassist.model.student.Email;
 import seedu.taassist.model.student.Name;
 import seedu.taassist.model.student.Phone;
 import seedu.taassist.model.student.Student;
-import seedu.taassist.model.student.StudentModuleData;
 
 /**
  * Contains utility methods for populating {@code TaAssist} with sample data.
