+++ src/main/java/seedu/taassist/storage/JsonAdaptedStudentModuleData.java
@@ -9,8 +9,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.taassist.commons.exceptions.IllegalValueException;
 import seedu.taassist.model.moduleclass.ModuleClass;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.session.SessionData;
-import seedu.taassist.model.student.StudentModuleData;
 
 /**
  * Json-friendly version of {@link StudentModuleData}.
