+++ src/main/java/seedu/taassist/storage/JsonSerializableTaAssist.java
@@ -12,9 +12,9 @@ import seedu.taassist.commons.exceptions.IllegalValueException;
 import seedu.taassist.model.ReadOnlyTaAssist;
 import seedu.taassist.model.TaAssist;
 import seedu.taassist.model.moduleclass.ModuleClass;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.session.SessionData;
 import seedu.taassist.model.student.Student;
-import seedu.taassist.model.student.StudentModuleData;
 
 /**
  * An Immutable TaAssist that is serializable to JSON format.
