+++ src/test/java/seedu/taassist/storage/JsonAdaptedStudentTest.java
@@ -11,11 +11,11 @@ import java.util.stream.Collectors;
 import org.junit.jupiter.api.Test;
 
 import seedu.taassist.commons.exceptions.IllegalValueException;
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.student.Address;
 import seedu.taassist.model.student.Email;
 import seedu.taassist.model.student.Name;
 import seedu.taassist.model.student.Phone;
-import seedu.taassist.model.student.StudentModuleData;
 
 public class JsonAdaptedStudentTest {
     private static final String INVALID_NAME = "R@chel";
