+++ src/test/java/seedu/taassist/testutil/StudentBuilder.java
@@ -3,12 +3,12 @@ package seedu.taassist.testutil;
 import java.util.ArrayList;
 import java.util.List;
 
+import seedu.taassist.model.moduleclass.StudentModuleData;
 import seedu.taassist.model.student.Address;
 import seedu.taassist.model.student.Email;
 import seedu.taassist.model.student.Name;
 import seedu.taassist.model.student.Phone;
 import seedu.taassist.model.student.Student;
-import seedu.taassist.model.student.StudentModuleData;
 import seedu.taassist.model.util.SampleDataUtil;
 
 /**
